requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pyaml
scikit-learn
nltk
scipy
spacy
gensim
mlflow
dagshub
joblib

-e .